explore: Add crossbeam channel
