+++ desktop-ui-foundation/Cargo.toml
@@ -12,3 +12,4 @@ fundsp = "0.20.0"
 miniquad = "0.4.8"
 glam = "0.30.5"
 midir = "0.10.2"
+ringbuf = "0.4.8"
